Fix deadlock when getting file object
Ensure date variable is updated when looking for log file fails. Otherwise, the program will deadlock when the program was started before the game and before UTC Midnight on a day when no log file does exist.

This bug was created in a previous implementation that required the date variable to be replaceable as parameter on function call of get_file_obj().